# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a online ordering  server.",
        "title": "Online Ordering System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v01/admin/login": {
            "post": {
                "description": "관리자가 로그인 하기 위한 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Login"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Admin"
                            }
                        }
                    }
                }
            }
        },
        "/v01/admin/menu/create": {
            "post": {
                "description": "메뉴를 만드는 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Create Menu"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "menuid",
                        "name": "menuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "imageurl",
                        "name": "imageurl",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quantity",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "price",
                        "name": "price",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "recommendation",
                        "name": "recommendation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "admin",
                        "name": "admin",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "score",
                        "name": "score",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "review",
                        "name": "review",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/v01/admin/menu/delete": {
            "post": {
                "description": "메뉴를 삭제하는 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete Menu"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "menuid",
                        "name": "menuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/v01/admin/menu/recom": {
            "post": {
                "description": "특정 메뉴를 추천해주는 함수다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommend Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "recommendation",
                        "name": "recommendation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/v01/admin/menu/status": {
            "get": {
                "description": "메뉴를 전부 가져오는 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get All Menu"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/v01/admin/register": {
            "post": {
                "description": "관리자를 등록하기 위한 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Register"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Admin"
                            }
                        }
                    }
                }
            }
        },
        "/v01/user/login": {
            "post": {
                "description": "로그인하는 함수",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    }
                }
            }
        },
        "/v01/user/menu/detail": {
            "get": {
                "description": "메뉴를 한개만 가져온다. 해당하는 메뉴만 가져온다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get  Menu Detail"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "menuid",
                        "name": "menuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/v01/user/order": {
            "post": {
                "description": "특정 메뉴를 주문한다. 주문데이터는 model.Order를 참고한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crete Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "menuname",
                        "name": "menuname",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "customer",
                        "name": "customer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "phonenumber",
                        "name": "phonenumber",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "quantity",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "paymentinformation",
                        "name": "paymentinformation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Order"
                            }
                        }
                    }
                }
            }
        },
        "/v01/user/register": {
            "post": {
                "description": "유저를 등록하는 함수다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Register user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Admin": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.Menu": {
            "type": "object",
            "properties": {
                "admin": {
                    "description": "관리자 이름",
                    "type": "string"
                },
                "create_time": {
                    "description": "생성시간",
                    "type": "string"
                },
                "imageurl": {
                    "description": "이미지 url",
                    "type": "string"
                },
                "menuid": {
                    "description": "메뉴 아이디, 추후 중복방지를 위해 둠",
                    "type": "integer"
                },
                "name": {
                    "description": "메뉴이름",
                    "type": "string"
                },
                "orders": {
                    "description": "추후 통계를 위해 개발 필요 오더스를 취합해서 평점 계산필요",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Order"
                    }
                },
                "price": {
                    "description": "메뉴가격",
                    "type": "integer"
                },
                "quantity": {
                    "description": "총 개수",
                    "type": "integer"
                },
                "recommendation": {
                    "description": "메뉴 추천/비추천",
                    "type": "boolean"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Review"
                    }
                },
                "score": {
                    "description": "점수",
                    "type": "integer"
                }
            }
        },
        "model.Order": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "주소",
                    "type": "string"
                },
                "customer": {
                    "description": "고객",
                    "type": "string"
                },
                "menuname": {
                    "description": "메뉴이름",
                    "type": "string"
                },
                "orderid": {
                    "type": "integer"
                },
                "paymentinformation": {
                    "description": "결제정보",
                    "type": "string"
                },
                "phonenumber": {
                    "description": "번호",
                    "type": "string"
                },
                "quantity": {
                    "description": "개수",
                    "type": "integer"
                },
                "rating": {
                    "description": "평점",
                    "type": "string"
                },
                "status": {
                    "description": "order status",
                    "type": "string"
                }
            }
        },
        "model.Review": {
            "type": "object",
            "properties": {
                "review_content": {
                    "description": "리뷰 콘텐츠",
                    "type": "string"
                },
                "review_title": {
                    "description": "리뷰 제목",
                    "type": "string"
                },
                "reviewid": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}